file: add opening
Add a custom File datatype and a mechanism for opening it. This allows
the basic std::fs::File system to be extended with shortcut methods an
d allows access to supplementary information without a Result needing
to handle a result. A temporary arg parsing mechanism has also been set
up.