id3: add apic resolution display
Add an extra formatter to the apic frame that will format the
resolution in the form WxH. This code uses some janky JPEG/PNG parsing
code to accomplish this so I don't have to roll an entire dependency
just to show image resolution. This could cause problems in the future,
but for now it just works.