id3: refactor tag creation
Refactor the tag creation code to parse all frames at once, isolate the
header code into it's own struct and parser, and to remove the need for
large reference chains that became a hassle to manage. This does
decrease the speed of the parser slightly, but it's still very efficent
compared to other parsers.