id3: fix crash with string data w/o nul
Fix a crash where get_nul_string will accidentally overread the given
slice if the slice does not have a NUL at the end. Instead,
get_nul_string will return the entire data slice as a string instead.
This should be a non-issue since terminating ID3 data with a NUL is a
must for a non-malformed file.