id3: refactor terminated string handling
The previous terminated string handling system was causing limitations
when used with UTF-16, so this commit refactors the system to base the
terminated string size of the raw byte size in it's own format. This
should resolve issues where the tag data would become extremely
misaligned because of UTF-16 sizes being interpreted as UTF-8.