id3: refactor frame parsing
Create a Id3Frame::parse method that all parsing code has been moved
to, mostly for better design. This also makes all new functions return
a default implementation of the frame with the specified FrameHeader
applied. String refactoring has also been refactored to be in line with
this new system.