id3v2: change frame ID error handling
Change the Frame ID error handling to simply panic on invalid input.
This is somewhat cleaner than the previous implementation and is more
in line with how such an error should be handled, so its okay. This
also adds some extra checks to ensure that TextFrames and UrlFrmes will
be encoded correctly.