id3v2: add chapter frame
Add a basic CHAP [chapter] frame implementation. This was the main
reason behind implementing a TagHeader reference in Frame::parse even
though it has no use outside of these two frames. If I were to store
a reference, it would create a lifecycle which would then make
downcasting impossible. If I were to copy, that would create needless
desync between the two or more instances. This is largely the best
solution.