id3v2: use get_terminated_string in text frames
Instead of the janky parse_text that didn't account for the many quirks
of nul-terminated strings, we instead just call get_terminated_string
repeatedly in text frames to get the full list of strings. This is much
more efficent as copying only occurs during decoding, while in the
original parser text would be copied in multiple places.