id3v2: add text rendering
Add a new render method onto the Frame trait that will return the byte
represention of whatever it is containing. The first frames to
implement rendering are TextFrame, UserTextFrame, and CreditsFrame.
Tests for these frames have also been slimmed down to an extent until
a more sustainable way to test nul-deliminited frames is added.