id3v2: seperate rendering and empty logic
Create a seperate function called is_empty that returns whether the
frame is considered empty by the spec. This is mostly for convienence
outside of a library, as frames that would be dropped can be rejected
if is_empty returns true. The render method no longer returns an
option with this change.