id3v2: improved fixed string handling
ID3v2 has a number of "Fixed String" types that arent handled well with
the normal, growable string. This commit improves the check on the OWNE
frame that checks if the purchase date string is valid, while also
adding a new language struct that properly guards against the language
fixed string being set to an invalid value.