id3v2: improve versioning
Make the versioning system more elegant by merging ID3v2.2 into the
main parsing code. This isn't ideal, as it introduces a variant in
Version that is theoretically never used and has to be sanitized when
used when creating a new tag, but it generally produces a better
implementation internally.