id3v2: imrpove unknown frame handling
The previous implementation of UnknownFrame technically made it a Frame
that could be added to a FrameMap if cloned. This fixes that issue by
making UnknownFrame a seperate struct that does not implement Frame.
As a side-effect, unknown frames can now contain ID3v2.2 frames that
could not be upgraded.